+++ CHANGELOG.md
@@ -9,7 +9,12 @@ Built-in operators:
 * import
 * include
 * reset
-* \# 
+* skip
+
+### v0.1.2 (2017-12-15)
+
+Bug fix. Ensure that the set of characters allowed in parameter and symbol names
+and the set of characters allowed as special characters are disjoint.
 
 ### v0.1.1 (2017-12-14)